unit tests finishing without error...

+++ ContigSorter.rb
@@ -5,7 +5,7 @@ class ContigSorter
         @file_count = 0
         @comp = comp
         @ex_id = ex_id
-        @contigs_filename = "#{dir}#{ex_id}"
+        @contigs_filename = "#{dir}/#{ex_id}"
     end
 
     def sort()
@@ -17,23 +17,28 @@ class ContigSorter
         return @file_count
     end
 
+    def compSize(e1,e2)
+        return @comp == 1 ? (e1 > e2) : (e1 < e2)
+    end
+
     def sortFile(filename)
         if(containsMinContigs(filename,2))
-            filename_L = "#{@ex_id}_L_#{getFileCount()}"
-            filename_R = "#{@ex_id}_R_#{getFileCount()}"
+            filename_L = "#{@contigs_filename}_L_#{getFileCount()}"
+            filename_R = "#{@contigs_filename}_R_#{getFileCount()}"
             contig_counter = 0
             while(nextContig = popTopContig(filename))
                 if(contig_counter % 2 == 0)
-                    append(filename_L)
+                    append(filename_L,nextContig)
                 else
-                    append(filename_R)
+                    append(filename_R,nextContig)
                 end
                 contig_counter += 1
             end
 
-            filename_L = sortFile(filename_L,getFileCount())
-            filename_R = sortFile(filename_R,getFileCount())
+            filename_L = sortFile(filename_L)
+            filename_R = sortFile(filename_R)
 
+            filename_M = "#{@contigs_filename}_M_#{getFileCount()}"
             return mergeFiles(filename_L,filename_R,filename_M)
         else
             return filename
@@ -45,7 +50,7 @@ class ContigSorter
             if(containsMinContigs(filename_L,1) && containsMinContigs(filename_R,1))
                 length_L = getTopContigLength(filename_L)
                 length_R = getTopContigLength(filename_R)
-                firstContig = eval("length_L #{@comp} length_R") ? popTopContig(filename_L) : popTopContig(filename_R)
+                firstContig = compSize(length_L,length_R) ? popTopContig(filename_L) : popTopContig(filename_R)
                 append(filename_M,firstContig)
             elsif(containsMinContigs(filename_L,1))
                 append(filename_M,popTopContig(filename_L))
@@ -60,14 +65,16 @@ class ContigSorter
     end
 
     def containsMinContigs(filename, min_limit)
-        filehandl = File.open(filename,"r")
         contig_count = 0
-        while(line = filehandl.gets && contig_count < min_limit)
-            if(line.match(/^>/))
-                contig_count += 1
+        if(File.exist?(filename))
+            filehandl = File.open(filename,"r")
+            while((line = filehandl.gets) && contig_count < min_limit)
+                if(line.match(/^>/))
+                    contig_count += 1
+                end
             end
+            filehandl.close
         end
-        filehandl.close
         return contig_count >= min_limit
     end
 
@@ -76,12 +83,15 @@ class ContigSorter
         length = 0
         contig_label = filehandl.gets
         if(contig_label.match(/^>/))
-            while(line = filehandl.gets && line.match(/(^[ATCGNatcgn]*\n?$)/))
+            while(line = filehandl.gets)
+               if(line.match(/(^[ATCGNatcgn]*\n?$)/))
                 length += line.count("ATCGNatcgn")
+               else
+                   break
+               end
             end
         else
-            puts "wtf malformed contig label '#{contig_label}' detected"
-            exit(1)
+            return nil
         end
         filehandl.close
         return length
@@ -91,26 +101,32 @@ class ContigSorter
         cp_filename = "#{filename}.tmp"
         filehandl = File.open(filename,"r")
         topContig = nil
-        contig_label = filehandl.gets
-        if(contig_label.match(/^>/))
-           topContig = contig_label
-           while(line = filehandl.gets && line.match(/(^[ATCGNatcgn]*\n?$)/))
-               topContig = "#{topContig}#{line}"
-           end
-           cp_filehandl = File.open(cp_filename,"w")
-           while(line = filehandl.gets)
-               cp_filehandl.puts(line)
-           end
-           filehandl.close
-           cp_filehandl.close
-           File.delete(filename)
-           File.rename(cp_filename,filename)
+        orig_contig_label = filehandl.gets
+        if(orig_contig_label.match(/^>/))
+            orig_topContig = orig_contig_label
+            cp_buf = nil
+            while(orig_line = filehandl.gets)
+                if(orig_line.match(/(^[ATCGNatcgn]*\n?$)/))
+                    orig_topContig = "#{orig_topContig}#{orig_line}"
+                else
+                    cp_buf = orig_line
+                    break
+                end
+            end
+            cp_filehandl = File.open(cp_filename,"w")
+            cp_filehandl.puts(cp_buf)
+            while(orig_line = filehandl.gets)
+                cp_filehandl.puts(orig_line)
+            end
+            filehandl.close
+            cp_filehandl.close
+            File.delete(filename)
+            File.rename(cp_filename,filename)
         else
             filehandl.close
-            puts "wtf malformed contig label '#{contig_label}' detected"
-            exit(1)
+            return nil
         end
-        return topContig
+        return orig_topContig
     end
 
     def append(filename, contig)
@@ -121,23 +137,39 @@ class ContigSorter
 
     def cpTopContigs(filename,cp_filename,lim)
         filehandl = File.open(filename,"r")
+        cp_filehandl = File.open(cp_filename,"w")
         count = 0
-        while(count < lim)
-            contig_label = filehandl.gets
-            if(contig_label.match(/^>/))
-                contig = contig_label
-                while(line = filehandl.gets && line.match(/(^[ATCGNatcgn]*\n?$)/))
-                    contig = "#{topContig}#{line}"
+        contig_label = filehandl.gets
+        contigs = ""
+        if(contig_label.match(/^>/))
+            contig_buf = contig_label
+            begin
+                contig_line = filehandl.gets
+                if(contig_line.nil?)
+                    contigs = "#{contigs}#{contig_buf}"
+                    contig_buf = contig_line
+                    break
+                elsif(contig_line.match(/^>/))
+                    count += 1
+                    contigs = "#{contigs}#{contig_buf}"
+                    contig_buf = contig_line
+                elsif(contig_line.match(/^[ATCGNatcgn]*\n?$/))
+                    contig_buf = "#{contig_buf}#{contig_line}"
+                else
+                    filehandl.close
+                    cp_filehandl.close
+                    puts "wtf malformed contig line '#{contig_line}' detected"
+                    exit(1)
                 end
-                cp_filehandl = File.open(cp_filename,"w")
-                cp_filehandl.puts(contig)
-                cp_filehandl.close
-            else
-                filehandl.close
-                puts "wtf malformed contig label '#{contig_label}' detected"
-                exit(1)
-            end
+            end while(count < lim)
+            cp_filehandl.puts(contigs)
+        else
+            filehandl.close
+            cp_filehandl.close
+            puts "wtf malformed contig label '#{contig_label}' detected"
+            exit(1)
         end
         filehandl.close
+        cp_filehandl.close
     end
 end
